Extract `Run` from `tmt.base.core` into `tmt.base.run`
Smaller files, smaller modules. Covered by pyright now.
