+++ tmt/__init__.py
@@ -20,8 +20,9 @@ __all__ = [
     'Tree',
 ]
 
-from tmt.base.core import Clean, Run, Status, Story, Test, Tree
+from tmt.base.core import Clean, Status, Story, Test, Tree
 from tmt.base.plan import Plan
+from tmt.base.run import Run
 from tmt.guest import Guest, GuestSsh
 from tmt.log import Logger
 from tmt.result import Result
