+++ tmt/base/plan.py
@@ -38,7 +38,6 @@ from tmt.base.core import (
     Core,
     FmfId,
     Links,
-    Run,
     Test,
     Tree,
     _RawAdjustRule,
@@ -46,6 +45,7 @@ from tmt.base.core import (
     _RawLink,
     expand_node_data,
 )
+from tmt.base.run import Run
 from tmt.container import SpecBasedContainer, container, field
 from tmt.lint import LinterOutcome, LinterReturn
 from tmt.utils import (
